SearchBoxコンポーネントにreact-hook-formを導入し、フォームの初期値を設定しました。また、SearchModalコンポーネントの初期パラメータを削除し、propsとしてtypeを受け取るように変更しました。
